Moved iOS MPPHandLandmark enum to MPPHandLandmarker.h
